emit special error on pool checkout timeout
